+++ src/main/java/WoWSSSC/model/WoWSAPI/shipprofile/profile/Fighters.java
@@ -9,16 +9,16 @@ import lombok.Data;
 @Data
 public class Fighters
 {
-    private long avg_damage;
+    private int avg_damage;
     private float cruise_speed;
     private long fighters_id;
     private String fighters_id_str;
     private float gunner_damage;
-    private long max_ammo;
+    private int max_ammo;
     private float max_health;
     private String name;
-    private long plane_level;
+    private int plane_level;
     private float prepare_time;
-    private long squadrons;
+    private int squadrons;
     private Fighters_Count_In_Squadron count_in_squadron;
 }
